// src/app/modules/students/student.interface.ts
import { Model, Types } from 'mongoose';


export type TUserName = {
  firstName: string;
  middleName: string;
  lastName: string;
}
export type TGuardian = {
  fatherName: string;
  fatherOccupation: string;
  fatherContactNo: string;
  motherName: string;
  motherOccupation: string;
  motherContactNo: string

};
export type TLocalGuardian = {
  name: string; 
  occupation: string;
  contactNo: string;
  address: string;

}
export type TStudent = {
  id: string
  user: Types.ObjectId;
  password: string
  name: TUserName
  gender: 'male' | 'female' | 'other'
  dateOfBirth: string
  email: string
  contactNo: string
  emergencyContact: string
  bloodGroups?: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-'
  presentAddress: string
  permanentAddress: string
  guardian: TGuardian
  localGuardian: TLocalGuardian
  profileImg?: string
  isDeleted: boolean
  admissionSemester:Types.ObjectId,
  academicDepartment:Types.ObjectId
}


export interface StudentModel extends Model<TStudent> {
  isUserExists(id: string): Promise<TStudent | null>
}